pyproject.toml
[project]
name = "adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.2",
    "docker>=7.1.0",
    "loguru>=0.7.3",
    "more-itertools>=10.8.0",
    "peft>=0.18.0",
    "polars>=1.36.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "swe-rex>=1.4.0",
    "tenacity",
    "transformers>=4.57.3",
    "trl>=0.26.2",
    "flash-attn",
    "oai-utils",
    "async-utils",
    "logfire>=4.16.0",
    "openhands-sdk>=1.6.0",
    "openhands-tools>=1.6.0",
    "openhands-workspace>=1.6.0",
    "openhands-agent-server>=1.6.0",
    "vllm>=0.11.0",
    "litellm>=1.80.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyrefly]
search-path = ["./oai_utils", "./async_utils"]
project-excludes = ["software-agent-sdk"]

[tool.uv]
no-binary-package = ["flash-attn"]
no-build-isolation-package = ["flash-attn"]

[tool.uv.workspace]
members = [
    "oai_utils",
    "async_utils",
]

[tool.uv.sources]
oai-utils = { workspace = true, editable = true }
async-utils = { workspace = true, editable = true }

